+++ recovery-gui/src/recovery_tool.py
@@ -19,6 +19,7 @@ import config  # noqa: E402
 from about_page import AboutPage  # noqa: E402
 from backup_page import BackupPage  # noqa: E402
 from restore_page import RestorePage  # noqa: E402
+from usb_page import USBPage  # noqa: E402
 
 
 class RecoveryWindow(Gtk.ApplicationWindow):
@@ -69,6 +70,7 @@ class RecoveryWindow(Gtk.ApplicationWindow):
 
         self.stack.add_titled(BackupPage(), "backup", "Backup")
         self.stack.add_titled(RestorePage(), "restore", "Restore")
+        self.stack.add_titled(USBPage(), "usb", "USB Writer")
         self.stack.add_titled(AboutPage(), "about", "About")